move check file from Writer to Path Handler
