+++ src/main/java/org/fundacionjala/at15/pokemon/io/Writer.java
@@ -9,7 +9,7 @@ import org.fundacionjala.at15.pokemon.Entity;
 public final class Writer {
     public static void writeToJson(Entity entity) {
         File path = new File(entity.getType().getPath());
-        checkFolder(path);
+        PathHandler.checkFolder(path);
         String fileName = entity.getId();
         String jsonString = new Gson().toJson(entity);
         try {
@@ -19,16 +19,6 @@ public final class Writer {
             file.close();
         } catch (IOException e) {
             e.printStackTrace(System.out);
-        }
-    }
-
-    public static void checkFolder(File path) {
-        if (path.exists()) {
-            System.out.println("Data saved!");
-        } else if (path.mkdirs()) {
-            System.out.println(path + " was created");
-        } else {
-            System.out.println(path + " was not created");
         }
     }
 }
